Removed Android dependency in OkLogInterceptorTest
